Completed express middleware
