+++ lib/client.js
@@ -40,12 +40,11 @@ module.exports = class Client {
 
   /**
    * Set the scopes for future requests
-   * @param {Scope} scopes
-   * @type {Array | String}
+   * @param {Scope || Array<Scope>} scopes
    */
   setScopes(...scopes) {
     if (scopes.length < 1) throw new Error('No scopes were provided.');
-    if (Array.isArray(scopes[0])) scopes = scopes[0];
+    if (Array.isArray(scopes[0])) scopes = scopes.flat();
 
     for (let scope of scopes) {
       if (!scopesList.includes(scope)) throw new Error('Invalid scope provided.');
