+++ lib/middlewares/index.js
@@ -1,0 +1,7 @@
+const verifySettings = require('./settings');
+const expressMiddleware = require('./express');
+
+module.exports = {
+  verifySettings,
+  expressMiddleware
+}
